requirements.txt
Django==5.2.4
Pillow==10.0.0
python-decouple==3.8
djangorestframework==3.14.0
django-cors-headers==4.3.1 